pkg/private/util: move ASList to tools/integration
