+++ pkg/private/util/BUILD.bazel
@@ -3,7 +3,6 @@ load("//tools/lint:go.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "aslist.go",
         "checksum.go",
         "docker.go",
         "duration.go",
@@ -15,17 +14,14 @@ go_library(
     importpath = "github.com/scionproto/scion/pkg/private/util",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/addr:go_default_library",
         "//pkg/private/common:go_default_library",
         "//pkg/private/serrors:go_default_library",
-        "@in_gopkg_yaml_v2//:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
     srcs = [
-        "aslist_test.go",
         "checksum_test.go",
         "duration_test.go",
         "export_test.go",
@@ -34,8 +30,6 @@ go_test(
     data = glob(["testdata/**"]),
     embed = [":go_default_library"],
     deps = [
-        "//pkg/addr:go_default_library",
-        "//pkg/private/xtest:go_default_library",
         "@com_github_stretchr_testify//assert:go_default_library",
         "@com_github_stretchr_testify//require:go_default_library",
     ],
