+++ private/path/combinator/BUILD.bazel
@@ -45,5 +45,6 @@ go_test(
         "//pkg/snet/path:go_default_library",
         "@com_github_golang_mock//gomock:go_default_library",
         "@com_github_stretchr_testify//assert:go_default_library",
+        "@com_github_stretchr_testify//require:go_default_library",
     ],
 )
