+++ private/revcache/revcachetest/BUILD.bazel
@@ -14,5 +14,6 @@ go_library(
         "//pkg/private/xtest:go_default_library",
         "//private/revcache:go_default_library",
         "@com_github_smartystreets_goconvey//convey:go_default_library",
+        "@com_github_stretchr_testify//require:go_default_library",
     ],
 )
