+++ tools/integration/BUILD.bazel
@@ -1,8 +1,9 @@
-load("//tools/lint:go.bzl", "go_library")
+load("//tools/lint:go.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
     srcs = [
+        "aslist.go",
         "binary.go",
         "cmd.go",
         "docker.go",
@@ -18,7 +19,6 @@ go_library(
         "//pkg/log:go_default_library",
         "//pkg/private/common:go_default_library",
         "//pkg/private/serrors:go_default_library",
-        "//pkg/private/util:go_default_library",
         "//pkg/snet:go_default_library",
         "//private/topology:go_default_library",
         "//tools/integration/progress:go_default_library",
@@ -27,3 +27,15 @@ go_library(
 )
 
 exports_files(["bin_wrapper.sh"])
+
+go_test(
+    name = "go_default_test",
+    srcs = ["aslist_test.go"],
+    data = glob(["testdata/**"]),
+    deps = [
+        ":go_default_library",
+        "//pkg/addr:go_default_library",
+        "//pkg/private/xtest:go_default_library",
+        "@com_github_stretchr_testify//assert:go_default_library",
+    ],
+)
