+++ tools/integration/aslist.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package util
+package integration
 
 import (
 	"os"
